PhysicalInventory class
This class inherits from Inventory and therefor will need to have some code uncommented out when the Inventory class is created.  This is a good example to follow for creating the NonPhysicalInventory class because it also inherits from Inventory class.

EndUser class

Has been tested and is running properly.

Game class

There is an issue with the date.  I imported the date and it fixed the errors inside the class.  It isn't running properly in the test run.